move ztf_join_gcn from online to fink_mm directory, remove the old offline
